move the generate_all function
